SomeApi.Api -> SomeApi, move BASEURL into appropriate Api classes
